Move message details to dedicated file
