+++ src/index.ts
@@ -3,6 +3,7 @@ export * from './formats';
 export * from './levels';
 export * from './log';
 export * from './log-by';
+export * from './log-details';
 export * from './log-level';
 export * from './log-message';
 export * from './log-recorder';
