+++ src/log-message.ts
@@ -2,6 +2,7 @@
  * @packageDocumentation
  * @module @run-z/log-z
  */
+import type { ZLogDetails } from './log-details';
 import type { ZLogLevel } from './log-level';
 import { ZLogMessageBuilder } from './log-message-builder.impl';
 import { ZLogMessageData__symbol } from './log-message-data.impl';
@@ -45,13 +46,6 @@ export interface ZLogMessage {
 }
 
 /**
- * Log message details map.
- *
- * The keys of this map are specific to application or log recorder implementation.
- */
-export type ZLogDetails = { readonly [key in string | symbol]?: any };
-
-/**
  * Builds a log message.
  *
  * Treats the first textual argument as {@link ZLogMessage.text message text}.
@@ -79,23 +73,6 @@ export function zlogMessage(level: ZLogLevel, ...args: any[]): ZLogMessage {
   builder.addAll(args);
 
   return builder.message();
-}
-
-/**
- * Builds a special value {@link zlogMessage treated} as additional {@link ZLogMessage.details message details}.
- *
- * The resulting value can be passed to {@link zlogMessage} function or to {@link ZLogger.log logger method} to add
- * details to logged message.
- *
- * @param details  Log message details to add.
- *
- * @returns A special value.
- */
-export function zlogDetails(details: ZLogDetails): unknown {
-  return {
-    [ZLogMessageData__symbol]: 'details',
-    details,
-  };
 }
 
 /**
