+++ src/logs/updated.log.spec.ts
@@ -1,7 +1,8 @@
 import { logZBy } from '../log-by';
+import { zlogDetails } from '../log-details';
 import { ZLogLevel } from '../log-level';
 import type { ZLogMessage } from '../log-message';
-import { zlogDetails, zlogMessage } from '../log-message';
+import { zlogMessage } from '../log-message';
 import type { ZLogRecorder } from '../log-recorder';
 import type { ZLogger } from '../logger';
 import { logZUpdated } from './updated.log';
