+++ src/logs/with-details.log.ts
@@ -2,7 +2,7 @@
  * @packageDocumentation
  * @module @run-z/log-z
  */
-import type { ZLogDetails } from '../log-message';
+import type { ZLogDetails } from '../log-details';
 import type { ZLogRecorder } from '../log-recorder';
 import { logZUpdated } from './updated.log';
