+++ src/node/to-stream.log.spec.ts
@@ -2,8 +2,9 @@ import * as os from 'os';
 import { levelZLogField, messageZLogField } from '../fields';
 import { textZLogFormatter } from '../formats';
 import { logZBy } from '../log-by';
+import { zlogDetails } from '../log-details';
 import { ZLogLevel } from '../log-level';
-import { zlogDetails, zlogMessage } from '../log-message';
+import { zlogMessage } from '../log-message';
 import { zlogDefer } from '../loggable';
 import { TestWritable } from '../spec';
 import { logZToStream } from './to-stream.log';
